Handle after payment properly
