+++ src/app/dashboard/(root)/page.tsx
@@ -1,10 +1,24 @@
-import { Suspense } from "react"
-import OnlinePaymentBlock from "~/components/payment/online/block"
-import { Skeleton } from "~/components/ui/skeleton"
+import { toast } from "~/components/ui/use-toast"
 import { api } from "~/trpc/server"
+import PaymentBlock from "./payment-block"
 
 export default async function Dashboard() {
   const user = await api.user.getCurrent.query()
+  const cards = await api.payment.getCards.query()
+
+  const handleAfterPayment = async (paymentID: string) => {
+    "use server"
+    await api.user.updateRole.mutate({
+      id: user.id,
+      role: "member",
+      paymentID,
+    })
+
+    toast({
+      title: "Successfully updated role",
+      description: "You are now a member of Coders for Causes",
+    })
+  }
 
   return (
     <>
@@ -33,9 +47,7 @@ export default async function Dashboard() {
                   </ul>
                 </div>
               </div>
-              <Suspense fallback={<Skeleton className="h-[320px] w-full" />}>
-                <OnlinePaymentBlock />
-              </Suspense>
+              <PaymentBlock user={user} cards={cards} afterPayment={handleAfterPayment} />
             </div>
           )}
         </div>
